+++ bin/hxltmdexml.py
@@ -28,12 +28,13 @@
 #       COMPANY:  EticaAI
 #       LICENSE:  Public Domain dedication OR Zero-Clause BSD
 #                 SPDX-License-Identifier: Unlicense OR 0BSD
-#       VERSION:  v0.7.0
+#       VERSION:  v0.9.0-rc.1
 #       CREATED:  2021-07-24 00:04 UTC v0.1.0 de hxl2example
 #      REVISION:  2021-07-24 17:49 UTC v0.2.0 hxltmdexml at least read XML
 #                 2021-07-25 23:28 UTC v0.5.0 XLIFF 1.2 and XLIFF 2.1 (MVP)
 #                 2021-07-28 22:01 UTC v0.7.0 TBX-IATE (MVP), TMX (MVP)
 #                 2021-11-06 21:19 UTC v0.8.0 Fix cli importing; more stable
+#                 -------------------- v0.9.0 To be released
 # ==============================================================================
 """hxltmdexml.py: Humanitarian Exchange Language Trānslātiōnem Memoriam de XML
 
@@ -189,7 +190,7 @@ import yaml
 # pip3 install langcodes
 # import langcodes
 
-__VERSION__ = "v0.7.1"
+__VERSION__ = "v0.9.0-rc.1"
 
 # _[eng-Latn]
 # Note: If you are doing a fork and making it public, please customize
@@ -274,30 +275,24 @@ STDIN = sys.stdin.buffer
 
 _HOME = str(Path.home())
 
-# TODO: clean up redundancy from hxlm/core/schema/urn/util.py
-HXLM_CONFIG_BASE = os.getenv(
-    'HXLM_CONFIG_BASE', _HOME + '/.config/hxlm')
-# ~/.config/hxlm/cor.hxltm.yml
+# cor.hxltm.215.ext:
+# 1. HXLTM_RUNNING_DIR ; defallo .
+# 2. HXLTM_DORMUM      ; defallo ~/.config/hxlm/cor.hxltm.215.ext
+# 3. HXLTM_SYSTEMA_DIR  ; defallo hxltmcli.py dir
 
 # _[eng-Latn]
-# This can be customized with enviroment variable HXLM_CONFIG_BASE
-#
-# Since hpd-toolchain is not a hard requeriment, we first try to load
-# hdp-toolchain lib, but if hxltmcli is a standalone script with
-# only libhxl, yaml, etc installed, we tolerate it
+# Environment variables HXLTM_DORMUM (for default search place for
+# configurations) HXLTM_TESTUM (for default place to search for test
+# files) can be manually configured.
 # [eng-Latn]_
-try:
-    from hxlm.core.constant import (
-        HXLM_ROOT,
-        HDATUM_EXEMPLUM
-    )
-    HXLTM_SCRIPT_DIR = HXLM_ROOT + '/core/bin'
-    HXLTM_TESTUM_BASIM_DEFALLO = str(HDATUM_EXEMPLUM).replace('file://', '')
-except ImportError:
-    HXLTM_SCRIPT_DIR = str(Path(__file__).parent.resolve())
-    HXLTM_TESTUM_BASIM_DEFALLO = str(Path(
-        HXLTM_SCRIPT_DIR + '/../../../testum/hxltm').resolve())
+HXLTM_DORMUM = os.getenv('HXLTM_DORMUM', _HOME + '/.config/hxltm')
+HXLTM_TESTUM = os.getenv(
+    'HXLTM_TESTUM', HXLTM_DORMUM + '/testum')
+# TODO: make HXLTM_TESTUM instalable as python package
 
+# systēma
+
+HXLTM_SYSTEMA_DIR = str(Path(__file__).parent.resolve())
 HXLTM_RUNNING_DIR = str(Path().resolve())
 
 # De Python Constants
@@ -2904,8 +2899,8 @@ class HXLTMUtil:
         # pylint: disable=using-constant-test
         if is_debug:
             print('load_hxltm_options')
-            print('HXLM_CONFIG_BASE', HXLM_CONFIG_BASE)
-            print('HXLTM_SCRIPT_DIR', HXLTM_SCRIPT_DIR)
+            print('HXLTM_DORMUM', HXLTM_DORMUM)
+            print('HXLTM_SYSTEMA_DIR', HXLTM_SYSTEMA_DIR)
             print('HXLTM_RUNNING_DIR', HXLTM_RUNNING_DIR)
 
         if custom_file_option is not None:
@@ -2919,13 +2914,25 @@ class HXLTMUtil:
             return HXLTMUtil._load_hxltm_options_file(
                 HXLTM_RUNNING_DIR + '/cor.hxltm.yml', is_debug)
 
-        if Path(HXLM_CONFIG_BASE + '/cor.hxltm.yml').exists():
+        if Path(HXLTM_RUNNING_DIR + '/cor.hxltm.yml').exists():
             return HXLTMUtil._load_hxltm_options_file(
-                HXLM_CONFIG_BASE + '/cor.hxltm.yml', is_debug)
+                HXLTM_RUNNING_DIR + '/cor.hxltm.215.yml', is_debug)
 
-        if Path(HXLTM_SCRIPT_DIR + '/cor.hxltm.yml').exists():
+        if Path(HXLTM_DORMUM + '/cor.hxltm.yml').exists():
             return HXLTMUtil._load_hxltm_options_file(
-                HXLTM_SCRIPT_DIR + '/cor.hxltm.yml', is_debug)
+                HXLTM_DORMUM + '/cor.hxltm.yml', is_debug)
+
+        if Path(HXLTM_DORMUM + '/cor.hxltm.215.yml').exists():
+            return HXLTMUtil._load_hxltm_options_file(
+                HXLTM_DORMUM + '/cor.hxltm.215.yml', is_debug)
+
+        if Path(HXLTM_SYSTEMA_DIR + '/cor.hxltm.yml').exists():
+            return HXLTMUtil._load_hxltm_options_file(
+                HXLTM_SYSTEMA_DIR + '/cor.hxltm.yml', is_debug)
+
+        if Path(HXLTM_SYSTEMA_DIR + '/cor.hxltm.215.yml').exists():
+            return HXLTMUtil._load_hxltm_options_file(
+                HXLTM_SYSTEMA_DIR + '/cor.hxltm.215.yml', is_debug)
         # print('oioioi')
 
         raise RuntimeError(
@@ -3030,7 +3037,7 @@ class HXLTMTestumAuxilium:
 
         _[eng-Latn]
         Note: this will try check if the enviroment variable
-        HXLTM_TESTUM_BASIM and only fallback to assume the entire
+        HXLTM_TESTUM and only fallback to assume the entire
         hdp-toolchain installation (or a fork from
         EticaAI/HXL-Data-Science-file-formats) on local disk.
 
@@ -3052,9 +3059,9 @@ class HXLTMTestumAuxilium:
 
         # if HDATUM_EXEMPLUM:
         # hxltmtestum = str(Path(
-        #     HXLTM_SCRIPT_DIR + '/../../../testum/hxltm').resolve())
+        #     HXLTM_SYSTEMA_DIR + '/../../../testum/hxltm').resolve())
 
-        praefixum = os.getenv('HXLTM_TESTUM_BASIM', HXLTM_TESTUM_BASIM_DEFALLO)
+        praefixum = os.getenv('HXLTM_TESTUM', HXLTM_TESTUM)
 
         if archivum:
             return praefixum + '/' + archivum
@@ -3077,8 +3084,8 @@ class HXLTMTestumAuxilium:
         if not os.path.isfile(exemplum_archivum):
             raise RuntimeError(
                 'HXLTMTestumAuxilium non-datum [{}]. '
-                'Requīsītum: dēfīnītiōnem HXLTM_TESTUM_BASIM. Exemplum:'
-                '> HXLTM_TESTUM_BASIM="/home/marcus/testum/" '
+                'Requīsītum: dēfīnītiōnem HXLTM_TESTUM. Exemplum:'
+                '> HXLTM_TESTUM="/home/marcus/testum/" '
                 'python3 -m doctest hxltmcli-de-marcus.py'
                 ' <'.format(exemplum_archivum))